+++ doc-site/pages/examples/viewmodel/fields/BooleanField/form-usage.tsx
@@ -4,7 +4,8 @@
  * This example shows the default widget that will be used in a [Form](doc:Form)
  * when using [@prestojs/ui-antd](/docs/ui-antd). See [getWidgetForField](doc:getWidgetForField).
  *
- * The default widget is [BooleanWidget](doc:BooleanWidget)
+ * The default widget is [BooleanWidget](doc:BooleanWidget) if `blank` is `false` otherwise it is
+ * [NullableBooleanWidget](doc:NullableBooleanWidget).
  *
  * `final-form` expects to be told when a field is a checkbox so it knows how to
  * handle undefined values. This example demonstrates this with the [Form.Item](doc:FormItem)
@@ -14,7 +15,6 @@
  */
 import { Form } from '@prestojs/final-form';
 import { AntdUiProvider, FormItemWrapper, FormWrapper, getWidgetForField } from '@prestojs/ui-antd';
-import { getId, hashId } from '@prestojs/util/';
 import { BooleanField, IntegerField, viewModelFactory } from '@prestojs/viewmodel';
 import { Button } from 'antd';
 import 'antd/dist/antd.min.css';
@@ -27,6 +27,7 @@ class ExampleModel extends viewModelFactory(
         receiveNewsletter: new BooleanField({
             label: 'Subscribe to newsletter',
             helpText: 'Check this to subscribe to our newsletter',
+            blank: true,
         }),
     },
     { pkFieldName: 'id' }