+++ js-packages/@prestojs/ui-antd/src/getWidgetForField.ts
@@ -5,7 +5,6 @@ import React from 'react';
 
 // RangeField is not included: its not meant to be used directly - TODO: mark it abstract?
 const mapping = new Map<string, FieldWidgetType<any, any>>([
-    ['BooleanField', React.lazy(() => import('./widgets/BooleanWidget'))],
     ['CharField', React.lazy(() => import('./widgets/CharWidget'))],
     ['CurrencyField', React.lazy(() => import('./widgets/CurrencyWidget'))],
     ['DateField', React.lazy(() => import('./widgets/DateWidget'))],
@@ -25,7 +24,6 @@ const mapping = new Map<string, FieldWidgetType<any, any>>([
     ['IPAddressField', React.lazy(() => import('./widgets/IPAddressWidget'))],
     ['JsonField', React.lazy(() => import('./widgets/JsonWidget'))],
     ['NumberField', React.lazy(() => import('./widgets/NumberWidget'))],
-    ['NullableBooleanField', React.lazy(() => import('./widgets/NullableBooleanWidget'))],
     ['PasswordField', React.lazy(() => import('./widgets/PasswordWidget'))],
     ['SlugField', React.lazy(() => import('./widgets/SlugWidget'))],
     ['TextField', React.lazy(() => import('./widgets/TextWidget'))],
@@ -39,6 +37,9 @@ const choicesMapping = new Map<string, FieldWidgetType<any, any>>([
     ['CharField', React.lazy(() => import('./widgets/CharChoicesWidget'))],
     ['IntegerField', React.lazy(() => import('./widgets/IntegerChoicesWidget'))],
 ]);
+
+const LazyBooleanWidget = React.lazy(() => import('./widgets/NullableBooleanWidget'));
+const BooleanWidget = React.lazy(() => import('./widgets/BooleanWidget'));
 
 function splitWidgetAndProps(
     maybeWidgetAndProps:
@@ -151,6 +152,12 @@ export default function getWidgetForField<
             if (_widget) {
                 return [_widget, { ...props, multiple: true }];
             }
+        }
+        if (fieldClassName === 'BooleanField' && !widget) {
+            if (_field.blank) {
+                return LazyBooleanWidget;
+            }
+            return BooleanWidget;
         }
         return widget;
     };